requirements_optimized.txt
# Flask Web Framework
flask==2.3.3
gunicorn==21.2.0

# Core Data Processing
pandas>=2.0.0,<3.0.0
numpy>=1.21.0,<2.0.0

# Google APIs & AI
google-api-python-client>=2.100.0,<3.0.0
google-auth-httplib2>=0.1.0,<1.0.0
google-auth-oauthlib>=1.0.0,<2.0.0
gspread>=5.10.0,<6.0.0
google-generativeai>=0.3.0,<1.0.0

# Environment & Configuration
python-dotenv>=1.0.0,<2.0.0
python-dateutil>=2.8.0,<3.0.0

# Image Processing & File Handling
Pillow>=10.0.0,<11.0.0

# Data Visualization
plotly>=5.15.0,<6.0.0

# Web Utilities
requests>=2.31.0
beautifulsoup4>=4.12.0,<5.0.0

# Database Support
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.0.0

# Web Framework Dependencies (auto-installed with Flask)
Werkzeug>=2.3.0,<3.0.0
Jinja2>=3.1.0,<4.0.0

# REMOVED UNUSED DEPENDENCIES:
# - openpyxl (Excel files not used in core system)
# - xlrd (Excel files not used in core system)
# - PyPDF2 (PDF processing not used)
# - crawl4ai (Market intelligence uses simpler approaches)
# - playwright (Heavy web automation not needed)
# - lxml (XML processing not used)